+++ src/js/control-view.js
@@ -59,6 +59,7 @@ module.exports = {
       position_msg:
       {x: false, y: false, z: false, a: false, b: false, c: false},
       axis_position: 0,
+      jog_adjust: 100,
       video_url: '',
       deleteGCode: false
     }