+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Buildbotics CNC Controller Firmware Changelog
 ## v0.3.25
  - Error on home if max-soft-limit <= min-soft-limit + 1. #139
  - Decrease boot time networking delay.
+ - Default to US keyboard layout. #145
 
 ## v0.3.24
  - Added unhome button on axis position popup.